refactor: simplify loading states and extract ProjectsComponent for better organization
